- Updates the location of the files.
