+++ packages/observability/src/store/store.ts
@@ -1,5 +1,4 @@
-export * from "./log-tailer";
-export * from "./observability-paths";
-export * from "./observability-run-manager";
-export * from "./observability-store-reader";
-export * from "./trace-deserializer";
+export * from "./managers/observability-run-manager";
+export * from "./readers/log-tailer";
+export * from "./readers/observability-store-reader";
+export * from "./readers/trace-deserializer";
